add Blog page v2,4
